WIP: Add support for pipelining requests.
Pipelining means that we can fully saturate the connection. This should be a better
solution than using multiple threads/connections, where we can never be sure how many
connections we should use. However, we should obviously test this in practice.

In the long term, we can hopefully replace the backend pool with a single upload queue
that is processed by two Trio tasks (one reading, one writing). In the common case (no
error) this should be very efficient. In case of errors, we simply re-submit all pending
requests (so they should be idempotent).

For reading, we should probabably have a separate queue.